Job-Hunting-Agent/requirements.txt
PyPDF2==3.0.1
python-dotenv==1.0.0
langgraph==0.0.69
langchain-google-genai==1.0.10
langchain-tavily==0.1.1
langchain-core==0.2.38
typing-extensions==4.9.0